decouple execution from run command compilation
