+++ src/index.tsx
@@ -25,8 +25,7 @@ import { runIcon, saveIcon } from '@jupyterlab/ui-components';
 import { addNodeActionCommands } from './commands/NodeActionCommands';
 import { Token } from '@lumino/coreutils';
 import { DockLayout } from '@lumino/widgets';
-import { xircuitsIcon, debuggerIcon, componentLibIcon, changeFavicon, xircuitsFaviconLink } from './ui-components/icons';
-import { startRunOutputStr } from './kernel/RunOutput';
+import { xircuitsIcon, componentLibIcon, changeFavicon, xircuitsFaviconLink } from './ui-components/icons';
 
 
 const FACTORY = 'Xircuits editor';
@@ -95,7 +94,6 @@ const xircuits: JupyterFrontEndPlugin<void> = {
     const tracker = new WidgetTracker<DocumentWidget>({
       namespace: "Xircuits Tracker"
     });
-
 
     // Add the widget to the tracker when it's created
     widgetFactory.widgetCreated.connect((sender, widget) => {
@@ -278,52 +276,18 @@ const xircuits: JupyterFrontEndPlugin<void> = {
       }
     };
 
-    function doRemoteRun(path: string, command: string, msg: string, url){
-
-      try {
-        let command_str = command + " " + path;
-        let code_str = "\nfrom subprocess import Popen, PIPE\n\n";
-
-        code_str += `command_str= "${command_str}"\n`;
-        code_str += "p=Popen(command_str, stdout=PIPE, stderr=PIPE, universal_newlines=True, shell=True)\n";
-        code_str += "print('Remote Execution in process...\\n')\n";
-        code_str += `print('Please go to ${url} for more details\\n')\n`;
-        code_str += `print('${msg}\\n')\n`;
-        code_str += "for line in p.stdout:\n";
-        code_str += "    " + "print(line.rstrip())\n\n";
-        code_str += "if p.returncode != 0:\n";
-        code_str += "    " + "print(p.stderr.read())";
-
-        return code_str;
-      } catch (e) {
-        console.log(e)
-      }
-    }
-
-    // Execute xircuits python script and display at output panel
+    // Execute command and display at output panel
     app.commands.addCommand(commandIDs.executeToOutputPanel, {
       execute: async args => {
         const xircuitsLogger = new Log(app);
-        const current_path = tracker.currentWidget.context.path;
-        const model_path = current_path.split(".xircuits")[0] + ".py";
-        const message = typeof args['runCommand'] === 'undefined' ? '' : (args['runCommand'] as string);
-        const runType = typeof args['runType'] === 'undefined' ? '' : (args['runType'] as string);
-        const config = typeof args['config'] === 'undefined' ? '' : (args['config'] as string);
 
         // Create the panel if it does not exist
         if (!outputPanel || outputPanel.isDisposed) {
           await createPanel();
         }
-
+    
         outputPanel.session.ready.then(async () => {
-          let code = startRunOutputStr();
-          if (runType == 'remote-run') {
-            // Run subprocess when run type is Remote Run
-            code += doRemoteRun(model_path, config['command'], config['msg'], config['url']);
-          } else {
-            code += "%run " + model_path + message
-          }
-
+          const code = args['code'] as string;
           outputPanel.execute(code, xircuitsLogger);
         });
       },
